refactor(task): rename file-manager.ts and merge schedule file modules
Implements #235 - Phase 4 of the unified file transfer system refactoring.

## Changes

### Renamed Task File Manager
- `task/file-manager.ts` → `task/task-files.ts`
- The old name was confusing as it suggests file transfer functionality
- The new name clearly indicates it manages task markdown files

### Merged Schedule File Modules
- Combined `schedule/schedule-file-scanner.ts` + `schedule/schedule-file-scanner.ts`
  → `schedule/schedule-watcher.ts`
- Shared frontmatter parsing logic (previously duplicated)
- Cleaner module organization with related functionality together

### Updated Imports
- All imports updated to use new file paths
- Test files updated accordingly

## Benefits

| Aspect | Before | After |
|--------|--------|-------|
| Task file manager name | Ambiguous (file-manager) | Clear (task-files) |
| Schedule modules | 2 separate files | 1 consolidated file |
| Frontmatter parsing | Duplicated | Shared |

## Test Results

```
Test Files  46 passed (46)
Tests       840 passed (840)
```
