+++ src/agents/evaluator.ts
@@ -22,7 +22,7 @@
 
 import { Config } from '../config/index.js';
 import type { AgentMessage, AgentInput } from '../types/agent.js';
-import { TaskFileManager } from '../task/file-manager.js';
+import { TaskFileManager } from '../task/task-files.js';
 import { BaseAgent, type BaseAgentConfig } from './base-agent.js';
 
 /**
